requirements.txt
jupyter
opencv-python
Pillow
dlib
imutils
